% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ch5.R
\docType{data}
\name{BB}
\alias{BB}
\title{Blood Alcohol Contents OSU}
\format{
A data frame with 16 rows and 2 variables:
\describe{
\item{Gender}{Reported as male or female}
\item{Weight}{Reported weight of subjects, in pounds}
\item{Beers}{number of 12 oz beers consumed}
\item{BAC}{blood alcohol content, in grams per decaliter}
\item{1st_Sober}{Before alcohol consumption sobriety rating}
\item{2nd_Sober}{After alcohol consumption sobriety rating}
}
}
\source{
Laura Dowling for data from her Ohio State class project, discussion based on https://bcs.whfreeman.com/WebPub/Statistics/shared_resources/EESEE/BloodAlcoholContent/index.html
}
\usage{
BB
}
\description{
A group of n=16 student volunteers at The Ohio State University drank a
number of beers and their BAC levels were measured. According to the Electronic
Encyclopedia of Statistcal Examples and Exercises, "The experiment took place
in February of 1986 at a student dormitory. Sixteen students volunteered to be
the subjects in the experiment. Each student blew into a breathalyzer ... to
indicate that his or her initial BAC was zero. The number (between 1 and 9) of
12 ounce beers to be drunk was assigned to each of the subjects by drawing
tickets from a bowl. Thirty minutes after consuming their final beer, students
had their BAC measured by a police officer of the OSU police department. The
officer also administered a road sobriety test before and after the alcohol consumption.
This involved performing four simple tasks, graded on a scale of 1 to 10 (ten being
a perfect rating), demonstrating coordination: balancing on one foot, touching
the tip of one’s nose with a forefinger, placing one’s head back with one’s eyes
closed, and walking heel to toe. The police officer was not aware of how much alcohol
each subject had consumed."
}
\keyword{datasets}
