% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ch5.R
\docType{data}
\name{BBA}
\alias{BBA}
\title{Blood Alcohol Contents Australia}
\format{
A data frame with 16 rows and 2 variables:
\describe{
\item{Gend}{Reported as male or female}
\item{Wgt}{Reported weight of subjects, in kg}
\item{Height}{Reported height, in cm}
\item{Age}{Age, in years}
\item{BAC}{blood alcohol content, in grams per decaliter (presumed units)}
\item{Wine}{Number of glasses of wine consumed in 45 minutes}
}
}
\source{
Prof. Kevin Donegan, Univ. of Western Sydney, discussion based on https://bcs.whfreeman.com/WebPub/Statistics/shared_resources/EESEE/BloodAlcoholContent/index.html
}
\usage{
BBA
}
\description{
University of Western Sydney
In October of 1993, 22 volunteers, mainly university staff and students,
attended a lunchtime “cocktail party.” Prior to the start of the experiment,
a police officer from the Drug and Alcohol Unit breath-tested all of the participants
using a breathalyzer. White wine (10\% alcohol/volume) was distributed in 120 ml
glasses, and each participant chose his or her own rate of drinking.
Snack food and water were available at all times.
After 45 minutes participants were told to stop drinking.
After waiting an additional 15 minutes each subject was breath-tested again
and the number of glasses consumed was recorded for all participants.#'
}
\keyword{datasets}
