% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagram_function.R
\name{catchNumeric}
\alias{catchNumeric}
\title{Catch numeric values stored as strings}
\usage{
catchNumeric(strList)
}
\arguments{
\item{strList}{A character vector}
}
\value{
A list of vectors, with each vector having the correct type, either character or numeric.
}
\description{
\code{catchNumeric()} takes a character list and determines if it is a valid numeric list.
Adapted from: https://stackoverflow.com/questions/24129124/how-to-determine-if-a-character-vector-is-a-valid-numeric-or-integer-vector
Post by Adriano Rivolli on Sep 21, 2017 at 16:30 accessed on Dec 10, 2024 at 4:30 pm
}
\examples{
charList <- c("A","B","C")
catchNumeric(charList)

numList <- c("1","2","3")
catchNumeric(numList)

mixList <- c("A","2","C","4")
catchNumeric(mixList)
}
