% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cites.R
\docType{data}
\name{cites}
\alias{cites}
\title{Data from journal citations data set from Niyazov et al. (2016)}
\format{
a \code{data.frame} with 31216 observations on 38 variables:
\describe{
\item{Div. Mathematical Sciences}{Indicator variable - remaining divisions similarly defined}
\item{citations}{Number of article citations}
\item{impact_factor_2012}{Impact factor of journal}
\item{title}{Title of article}
\item{year}{Year of publication}
\item{avail_online_score}{Not used}
\item{date_collected}{Date information collected on article}
\item{source}{On Academia.edu (on) or not (off)}
\item{onacad}{Binary for on Academia.edu or not}
\item{age}{age of article in years (see exploration below)}
\item{online}{Binary for posted online somewhere}
\item{journal_title}{Journal title}
\item{topics}{Topic area}
\item{divisions}{Division of journal used to create first 23 indicators}
\item{impact_factor}{Impact factor of journal}
\item{ArticleID}{Unique ID for each article}
}
}
\source{
devtools::install_github('polynumeral/academia-citations/acadcites')
}
\usage{
cites
}
\description{
These data are on characteristics of journal articles including numbers of citations, originally from the acadcites R package
}
\examples{
 # Data wrangling and sampling
data(cites)
library(tidyverse)
set.seed(406)
cites_larger <- cites \%>\%
  group_by(journal_title) \%>\%
  mutate(totalcountsbyjournal = n()) \%>\% #add count of articles by journal
  dplyr::filter(totalcountsbyjournal > 10) #filter to only journals with >10 articles (note impacts on SOI)

cites_strat <- cites_larger \%>\%
  slice_sample(prop = 0.5) \%>\% #Sample 50\% of articles in each selected journal
  ungroup() \%>\% #Remove group_by information from data set
  dplyr::select(-totalcountsbyjournal) #Remove temporary variable

cites_strat <- cites_strat \%>\% mutate(
  onacad = factor(onacad),
  onacad = fct_recode(onacad, #Change labels to be more explicit
                      "NotOnAcad" = "0",
                      "OnAcad" = "1"),
  online = factor(online),
  online = fct_recode(online,
                      "NotOnline" = "0",
                      "Online" = "1"))

#Just focus on Div. Chemical Sciences, Div. Environmental Sciences, Div. Biological Sciences, Div. Engineering, Div. Medical and Health Sciences as these are most common categories
cites_strat <- cites_strat \%>\% mutate(
  journal_title = factor(journal_title),
  divisions = factor(divisions),
  yearssincepub = 2014 - year) \%>\%
  rename( ChemSci = 'Div. Chemical Sciences',
          EnvSci = 'Div. Environmental Sciences',
          BioSci = 'Div. Biological Sciences',
          Eng = 'Div. Engineering',
          MedHealthSci = 'Div. Medical and Health Sciences') \%>\%
          mutate( ChemSci = factor(ChemSci),
                  EnvSci = factor(EnvSci),
                  BioSci = factor(BioSci),
                  Eng = factor(Eng),
                  MedHealthSci = factor(MedHealthSci),) \%>\%
     dplyr::select(-starts_with("Div.")) #Removing unused division variables (remove code to explore your favorite discipline, but note that these are more rare in data set)
library("yarrr")
pirateplot(log1p(citations) ~ onacad, data = cites_strat,
  main = "Pirate plot of log-publications", inf.method = "ci",
  inf.disp = "line",
  pal = "southpark") #Bar is mean with 95\% CI
pirateplot(log1p(citations) ~ onacad+online, data = cites_strat,
   main = "Pirate plot of log-publications", inf.method = "ci",
   inf.disp = "line",
   pal = "southpark")
cites_strat <- cites_strat \%>\% mutate(logpub = log1p(citations))
  enhanced_stripchart(logpub ~ onacad, data = cites_strat)
cites_strat \%>\% mutate(logpub = log1p(citations)) \%>\%
  ggintplot(response = "logpub", groupvars = c("onacad", "online"), array = T)
}
\references{
Niyazov, Y., Vogel, C., Price, R., Lund, B., Judd, B., Akil, A., Mortonson, M., Schwartzman, J., and Shron, M. (2016). Open Access Meets Discoverability: Citations to Articles Posted to Academia.edu. \emph{PLoS ONE} 11(2): e0148257. doi:10.1371/journal.pone.0148257
}
\keyword{citations}
\keyword{data}
\keyword{journal}
