% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced-stripchart.R
\name{enhanced_stripchart}
\alias{enhanced_stripchart}
\title{Enhanced stripchart with violins and confidence intervals}
\usage{
enhanced_stripchart(
  data = NULL,
  formula = NULL,
  na.rm = TRUE,
  conf.level = 0.95,
  .drop = TRUE,
  ptalpha = 0.3,
  jitterwidth = 0.1,
  jitterheight = 0,
  ylab = NULL,
  xangle = 45,
  xvjust = 1,
  xhjust = 1,
  CLD = F,
  CLDvlocation = 0,
  CLDcolor = "red"
)
}
\arguments{
\item{data}{data frame or tibble name that contains the variables in the formula}

\item{formula}{a formula object such as y ~ x, where is y is quantitative and x is a variable that you want to group y based on}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{.drop}{option for plyr, mostly not used}

\item{ptalpha}{alpha value for the points, set to 0 to hide points and just view means, CIs, and violins}

\item{jitterwidth}{control jitter amount in horizontal, defaults to 0.1}

\item{jitterheight}{control jitter amount in vertical, defaults to 0}

\item{xangle}{angle for text on x-axis, defaults to 45}

\item{xvjust}{vertical location for the x-axis text, defaults to 1}

\item{xhjust}{horizontal location for the x-axis text, defaults to 1}

\item{CLD}{Add compact letter display using Tukey's HSD, default is FALSE}

\item{CLDvlocation}{yaxis value for CLD letters, default is 0.}

\item{CLDcolor}{defaults to red for adding CLD letters}

\item{conf.level:}{the proportion confidence level of the confidence interval (default is 0.95)}
}
\description{
This function allows you to use a formula interface to generate a plot much like a pirateplot using ggplot2.
}
\details{
It checks for missing data, converts the x variable to a factor, checks for smaller group sizes, and generates frequentist intervals for each level of the factor provided.

Function for making enhanced stripcharts with means and confidence intervals for each combination of groups. With option to add Tukey's HSD compact letter display.
}
\examples{
# must have the carData package installed
data(TitanicSurvival, package = "carData")
TitanicSurvival \%>\% enhanced_stripchart(age ~ passengerClass)
data(ToothGrowth) #Quantitative predictor with 3 levels, converted to factor
enhanced_stripchart(len ~ dose, data = ToothGrowth, ptalpha = 0.05,
  jitterwidth = 0.1, jitterheight = 0, CLD = T) +
  labs(title = "Plot of odontoblast cell lengths \n by dosage level",
    x = "Dosage of vitamin C",
    y = "Cell length (microns)")

}
