% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagram_function.R
\name{getFixLevel}
\alias{getFixLevel}
\title{Identify the fixed effect model structure level}
\usage{
getFixLevel(lme_model, fixedCall, randomCall)
}
\arguments{
\item{lme_model}{A \code{nlme} lme model object inherited from \code{model_diagram()}}

\item{fixedCall}{The fixedCall object from an lme model or parsed text
containing fixed variable names from a merMod object}

\item{randomCall}{The randomCall object from an lme model or parsed text
containing the random effect structure from a merMod object}
}
\value{
The random effect at which the fixed effect is placed at for degrees
of freedom calculations.
}
\description{
\code{getFixLevel()} uses the degrees of freedom information from \code{lme()}
to identify the correct placement of each fixed effect and interaction term
in a model structure diagram for hierarchical mixed effects models.
}
