% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdredgeplot.R
\name{ggdredgeplot}
\alias{ggdredgeplot}
\title{Number line plot of dredge IC model results}
\usage{
ggdredgeplot(
  data,
  deltasubset = NA,
  colors = cet_pal(6, "l3"),
  name = "Delta IC",
  max.overlaps = 200
)
}
\arguments{
\item{data}{Dredge model selection table output}

\item{deltasubset}{default is NA to show all results, cuts results to numeric delta IC value}

\item{colors}{default cet_pal(6, "l3"), colors for four binned delta IC values}

\item{name}{label for the x-axis number line in the plot}

\item{max.overlaps}{Number of overlapping labels for models, defaults to 200}
}
\description{
This function provides a ggplot2 of dredge model information-criteria results,
from the dredge package from the MuMIn R package, by Mark Greenwood
}
\details{
Function for visual display of IC dredge model selection results.
}
\examples{
library(dplyr)
library(ggrepel)
library(cetcolor)
data(powerplantoutput)
set.seed(406)
training_data <- slice_sample(powerplantoutput, prop = 0.7)
m1 <- lm(PE ~ (AT + V + AP + RH)^2, data = training_data)
library(MuMIn)
options(na.action = "na.fail")
dd <- dredge(m1, rank = "BIC")
g1 <- ggdredgeplot(data = dd) + labs(title = "BIC results for all models")
g2 <- ggdredgeplot(data = dd, deltasubset = )
library(patchwork)
g1 / g2 + plot_annotation((tag_levels = "a"))
}
