% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdredgeplotly.R
\name{ggdredgeplotly}
\alias{ggdredgeplotly}
\title{Interactive number line plot of dredge IC model results}
\usage{
ggdredgeplotly(
  data,
  deltasubset = max(data$delta) + 1,
  colors = cet_pal(6, "l3"),
  name = "Delta IC"
)
}
\arguments{
\item{data}{Dredge model selection table output}

\item{deltasubset}{default is max delta + 1 to show all results, cuts results to numeric delta IC value}

\item{colors}{default cet_pal(6, "l3"), colors for four binned delta IC values}

\item{name}{label for the x-axis number line in the plot}
}
\description{
This function provides a plotly display of dredge model information-criteria results,
from the dredge package from the MuMIn R package, by Mark Greenwood
}
\details{
Function for visual display of IC dredge model selection results.
}
\examples{
library(dplyr)
library(ggrepel)
library(cetcolor)
data(powerplantoutput)
set.seed(406)
training_data <- slice_sample(powerplantoutput, prop = 0.7)
m1 <- lm(PE ~ (AT + V + AP + RH)^2, data = training_data)
library(MuMIn)
options(na.action = "na.fail")
dd <- dredge(m1, rank = "BIC")
ggdredgeplotly(dd)

}
