% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagram_function.R
\name{inner_perc_table_R}
\alias{inner_perc_table_R}
\title{Table of inner group percentages for each fixed effect variable}
\usage{
inner_perc_table_R(X, grps, p = NULL, Q = NULL)
}
\arguments{
\item{X}{Matrix or data frame from lme object}

\item{grps}{Matrix or data frame where each column is a grouping variable}

\item{p}{If \code{NULL} the number of columns of \code{X}, else a pre-specified number
of fixed effects.}

\item{Q}{If \code{NULL} the number of columns in \code{grps}, else a pre-specified number
of random effects or groups.}
}
\value{
A matrix of inner group proportions.
}
\description{
\code{inner_perc_table_R()} Calculates the inner group percentage for each fixed
effect variable.
}
\details{
Translated \code{inner_perc_table()} from nlme's nlmefit.c file (https://github.com/cran/nlme/blob/master/src/nlmefit.c)
Code translation assistance provided by Anthropic's Claude 3.5 Sonnet (2024 version) on December 17, 2024.
}
\examples{
\dontrun{inner_perc_table_R(X, grps)}
}
