% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagram_function.R
\name{md}
\alias{md_color}
\alias{md_fill}
\alias{md_fontColor}
\title{Diagram elements}
\usage{
md_color(diagram = "gray25", random = "gray25", fixed = "gray25")

md_fill(diagram = "aliceblue", random = "aliceblue", fixed = "darkseagreen1")

md_fontColor(diagram = "black", random = "black", fixed = "black")
}
\arguments{
\item{diagram}{Specifies the outline color, fill color, or font color
for the elements in the measurement diagram circles (hierarchical diagram).
Default is \code{"gray25"} for (\code{\link[=md_color]{md_color()}}), \code{"aliceblue"} for (\code{\link[=md_fill]{md_fill()}}),
and \code{"black"} for (\code{\link[=md_fontColor]{md_fontColor()}}).}

\item{random}{Specifies the outline color, fill color, or font color
for the elements in the random effect variable boxes.
Default is \code{"gray25"} for (\code{\link[=md_color]{md_color()}}), \code{"aliceblue"} for (\code{\link[=md_fill]{md_fill()}}),
and \code{"black"} for (\code{\link[=md_fontColor]{md_fontColor()}}).}

\item{fixed}{Specifies the outline color, fill color, or font color
for the elements in the fixed effect variable boxes.
Default is \code{"gray25"} for (\code{\link[=md_color]{md_color()}}), \code{"darkseagreen1"} for (\code{\link[=md_fill]{md_fill()}}),
and \code{"black"} for (\code{\link[=md_fontColor]{md_fontColor()}}).}
}
\description{
Similar to \link{ggplot2}'s \link{theme} system, the \code{md_} functions
specify the display of how the node components of the diagram are drawn.
\itemize{
\item \code{md_color()}: border of the node.
\item \code{md_fill()}: fill color of the node.
\item \code{md_fontColor()}: font color of the text in the nodes.
}
}
\examples{
library(tidyverse)
library(DiagrammeR)
# merMod object example
library(lme4)
library(nlme)

sleepstudy_lmer <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
summary(sleepstudy_lmer)
model_diagram(sleepstudy_lmer,
  nodeColors = md_color(diagram="steelblue", random="steelblue", fixed="darkolivegreen1"),
  nodeFillColors = md_fill(diagram="aliceblue", random="aliceblue", fixed="darkseagreen1"),
  nodeFontColors = md_fontColor(diagram="black", random="blue", fixed="black"))

}
