% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concise.R
\name{print.summary.lm}
\alias{print.summary.lm}
\title{Simplifies the standard linear model summary output}
\usage{
\method{print}{summary.lm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  concise = TRUE,
  ...
)
}
\arguments{
\item{x}{model to be summarized}

\item{digits}{number of digits to show, defaults to max(3L, getOption("digits") - 3L)}

\item{signif.stars}{uses current option by default}

\item{concise}{defaults to TRUE, turn off to not shorten output}
}
\description{
Simplifies the standard linear model summary output
}
\details{
Reduces output to just needed aspects when lm model summary is printed.
}
\examples{
x <- sort(rep(1:5,10)); table(x)
y <- .5+.5*scale(x)-.5*(scale(x)^2)-.1*(scale(x)^3)+rnorm(50,sd=.5);
d1 <- data.frame(x, y) \%>\% mutate(xf = factor(x))
catmodel <- lm(y~xf, data = d1)
print(summary(catmodel))
}
