% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rat_nicotine2.R
\docType{data}
\name{rat_nicotine2}
\alias{rat_nicotine2}
\title{Data from rat tibia data sets from Barra et al. (2025) - Dataset 2 of 3}
\format{
a \code{data.frame} with 56 observations on 7 variables:
\describe{
\item{SUBSTANCE}{treatment substance ('SS': saline solution; 'NIC': nicotine)}
\item{SOLUTION}{'VG': vegetable glycerin; 'Q10': coenzyme Q10}
\item{EUTHANASIA}{timing of euthanasia (days post-treatment)}
\item{SAMPLE}{sample number}
\item{MAX_STRENGTH}{maximum tibial strength (N)}
\item{MAX_RESISTANCE}{maximum tibial resistance (N)}
\item{MODULUS_ELASTICITY}{modulus of elasticity (MPa)}
}
}
\source{
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0315462}, data for R package prepared by Amelia Pomazal
}
\usage{
rat_nicotine2
}
\description{
These data contain measurements related to bone health and gastrointestinal conditions in rats subjected to nicotine exposure and coenzyme Q10 supplementation. See original paper for use of these data in figures 9, 10, and 11.
}
\examples{

library(tidyverse)
rat_nicotine2 <- rat_nicotine2 \%>\%
  mutate(GROUP = paste(SUBSTANCE, SOLUTION, EUTHANASIA, sep = "_"),
     GROUP = factor(GROUP))
     
# View two-way interactions between the treatment levels
 ggintplot(response = "MAX_RESISTANCE", groupvars = c("SUBSTANCE", "SOLUTION", "EUTHANASIA"), data = rat_nicotine2)

# Enhanced stripchart of tibial strength across groups
  enhanced_stripchart(data = rat_nicotine2, MAX_STRENGTH ~ GROUP) +
     labs(title = "Tibial Weight Across Treatment Groups",
       x = "Experimental Group",
       y = "Maximum Tibial Strength (N)")
  
  
}
\references{
Barra, R. H. D., Piovezan, B. R., Matheus, H. R., Vitória, O. A. P., de Abreu Furquim, E. M., Fiorin, L. G., Santos, E. O., & de Almeida, J. M. (2025). Effect of coenzyme Q10 on tibial fracture resistance in nicotine-exposed rats. \emph{PLoS ONE, 20}(1), e0315462. https://doi.org/10.1371/journal.pone.0315462
}
\keyword{Q10}
\keyword{bone}
\keyword{coenzyme}
\keyword{data}
\keyword{health}
\keyword{nicotine}
\keyword{rats}
