% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vampirebatmetab.R
\docType{data}
\name{vampirebatmetab}
\alias{vampirebatmetab}
\title{Data from a repeated measures study on vampire bats}
\format{
a list of data.frames with many observations on multiple variables from 24 subjects:
\describe{
\item{bat_id}{Bat ID from 1 to 24 in demographics data set}
\item{date_of_capture}{Date of capture}
\item{mass_g}{Mass of bat in grams}
\item{sex}{sex of bats, male or female}
\item{treatment}{Additive contained in blood fed to bats, Control, Glycine, Leucine}
\item{blood_consumed_ml}{Blood consumed by bat in ml}
\item{Bat ID}{Bat ID code in repeated measures data set}
\item{Treatment}{Additive contained in blood fed to bats, Control, Glycine, Leucine}
\item{Mass (g)}{Mass in grams of bats}
\item{VO2_XXX_RepX}{VO2 measurement for XXX speed and Rep 1,2,3}
\item{VCO2_XXX_RepX}{VCO2 measurement for XXX speed and Rep 1,2,3}
}
}
\source{
Rossi, Giulia S; Welch, Kenneth C (2024). Supplementary material from "Vampire bats rapidly fuel running with essential or non-essential amino acids from a blood meal". The Royal Society. Collection. https://doi.org/10.6084/m9.figshare.c.7524683.v1
}
\usage{
vampirebatmetab
}
\description{
Results of a study of 24 bats on metabolism based on different diet types.
}
\details{
Data contains static demographic file in slot 1 (\code{bat_demo}) and wide file containing
repeated measures in slot 2 (\code{bat_metab}).

Code demonstrates \code{left_joining} and then pivoting the data for VCO2, similar
code can be used for VO2 measurements - see original papers for more details
}
\examples{
library(ggthemes)
library(dplyr)
library(ggplot2)
library(viridis)
library(janitor)
library(lmerTest)
library(effects)
data(vampirebatmetab)
vampirebatmetab$bat_demo \%>\% enhanced_stripchart(mass_g ~ treatment)
anova(lm(mass_g ~ treatment, data = vampirebatmetab$bat_demo))

VCO2only <- vampirebatmetab$bat_metab \%>\%  dplyr::select('Bat ID', Treatment, starts_with("VCO2"))
joineddataC <- VCO2only \%>\% left_join(vampirebatmetab$bat_demo, c('Bat ID' = 'bat_id'))
longVCO2 <- joineddataC \%>\% pivot_longer(cols = 3:11, values_to = "VCO2")
longVCO2 <- longVCO2 \%>\%
  mutate(Speed = str_sub(name,6,6),
           Rep = str_sub(name, nchar(name), nchar(name))
           ) \%>\% clean_names()

longVCO2select <- longVCO2 \%>\% dplyr::select(bat_id, treatment, mass_g, sex, blood_consumed_m_l, name, vco2, speed, rep)
VCO2R <- longVCO2select \%>\% drop_na(vco2)
VCO2R <- VCO2R \%>\% mutate(vco2perg = vco2/mass_g)
VCO2R <- VCO2R \%>\% mutate(Speed = factor(speed),
                          Speed1 = fct_reorder(Speed, vco2, .fun = mean),
                          Speed2 = as.numeric(Speed1)*10,
                          rep = factor(rep))

 VCO2R <- VCO2R \%>\% mutate(logvco2perg = log(vco2perg))
 VCO2R \%>\% ggplot(aes(x = Speed1, y = logvco2perg, group = bat_id, color = bat_id)) +
            geom_point() +
            geom_line() +
            facet_wrap(~treatment + rep)
ggintplot(response = "vco2perg", groupvars = c("Speed1", "treatment", "rep"), data = VCO2R, array = F)
 lmerFull1 <- lmer(vco2perg  ~ Speed1 * treatment * rep + sex + mass_g + (1|bat_id/rep), data = VCO2R)
 anova(lmerFull1)
}
\references{
Rossi GS, Welch KC. 2024 Vampire bats rapidly fuel running with essential or non-essential amino acids from a blood meal. Biology Letters. 20: 20240453. https://doi.org/10.1098/rsbl.2024.0453

Rossi Giulia S. and Welch Kenneth C. 2025 Correction: ‘Vampire bats rapidly fuel running with essential or non-essential amino acids from a blood meal’ (2024), by Rossi and WelchBiol. Lett.2120250088 http://doi.org/10.1098/rsbl.2025.0088
}
\keyword{bats}
\keyword{data}
\keyword{measures}
\keyword{metabolism}
\keyword{mixed}
\keyword{models}
\keyword{repeated}
