% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagram_function.R
\name{model_diagram}
\alias{model_diagram}
\title{Hierarchical model diagramming}
\usage{
model_diagram(
  this_model,
  this_file_path = NULL,
  this_file_type = "PNG",
  widthVal = 800,
  heightVal = 1600,
  includeSizes = TRUE
)
}
\arguments{
\item{this_model}{Input model. Either a lme or merMod object with a nested
random effects structure.}

\item{this_file_path}{Optional. Path to a location to export the diagram.
Default is \code{NULL}.}

\item{this_file_type}{Optional. File type to export the diagram.
Default is \code{"PNG"}.}

\item{widthVal}{Optional. Width of diagram in pixels. Default is \code{800}.}

\item{heightVal}{Optional. Height of diagram in pixels. Default is \code{1600}.}

\item{includeSizes}{Optional. Include group sizes in random effect labels.
Default is \code{TRUE}.}
}
\value{
A DiagrammeR object.
}
\description{
\code{model_diagram()} takes a hierarchical nested model and returns a DiagrammeR
object visualizing the fixed and random effects structure.
}
\examples{
library(tidyverse)
library(DiagrammeR)
# merMod object example
library(lme4)
library(nlme)

sleepstudy_lmer <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
summary(sleepstudy_lmer)
model_diagram(sleepstudy_lmer)

# lme object example

sleepstudy_lme <- lme(Reaction ~ Days, random=~Days|Subject, data=sleepstudy)
summary(sleepstudy_lme)
model_diagram(sleepstudy_lme)
}
