% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-TendonData.R
\docType{data}
\name{TendonData}
\alias{TendonData}
\title{Data on Achilles Tendons and other aspects of subjects}
\format{
a \code{data.frame} with 182 observations on 15 variables:
\describe{
\item{Subject ID}{identifier includes site info}
\item{Side}{Side of body}
\item{PSFR}{intra-tendinous morphology, units: mm^-1}
\item{HeelRaise}{Count of heel raise endurance test}
\item{KneeWall}{knee-to-wall composite dorsiflexion test, units?}
\item{CurrPain (Y=1)}{current pain}
\item{Neovascularization (Doppler) (Y=1)}{?}
\item{VISA-A}{Victorian Institute of Sport Assessment—Achilles score(?)}
\item{Age}{Age in years}
\item{Sex  (M=1)}{Male is 1, female is 0}
\item{YearsRunning}{Years running in years}
\item{BMI}{Body Mass Index}
\item{Waist to Hip}{Waist to hip ratio}
\item{CSA}{tendon gross morphology (average of three scans), units?}
\item{Hx Pain (Y=1)}{history of ankle/achilles pain}

}
}
\source{
Version 4 of the posted data on figshare is the correct version: https://figshare.com/articles/dataset/FigShare_Data_xlsx/8313386
}
\usage{
TendonData
}
\description{
Updated 2021 version of data set posted to accompany
Ho et al.’s “Factors related to intra-tendinous
morphology of Achilles tendon in runners” from figshare.
}
\details{
Note that data are as posted but variable names are manipulated in help to provide
variable names and coding that are more compatible with R.

"Subject ID" = identifier includes site info (two observations per subject typically)
"Side" = Leg levels of Left/Right
"PSFR" = See paper, main response variable related to intra-tendinous morphology, units: mm^-1, average of 3 scans
"HeelRaise" = Count of heel raise endurance test
"KneeWall" = knee-to-wall composite dorsiflexion test (a weight-bearing lunge test of composite ankle dorsiflexion range of motion, as described by Hoch and Mckeon)
"CurrPain (Y=1)" = current pain (in either/measured ankle? - check)
"Neovascularization (Doppler) (Y=1)" = ?
"VISA-A" = Victorian Institute of Sport Assessment—Achilles score(?)
"Age" = Age in years
"Sex  (M=1)" (53 males, 38 females in study)
"YearsRunning" = Years running (years)
"BMI"
"Waist to Hip" = waist to hip ratio
"CSA" = tendon gross morphology (average of three scans)
"Hx Pain (Y=1)" = history of ankle/achilles pain
}
\examples{
library(tidyverse)
library(catstats2)
data(TendonData)
  TendonData <- TendonData \%>\%
    dplyr::rename(SubjectID = 'Subject ID',
         Sex = 'Sex  (M=1)',
         CurrPain = 'CurrPain (Y=1)',
         WaisttoHip = 'Waist to Hip',
         VISAA = 'VISA-A',
         Neovascularization = 'Neovascularization (Doppler) (Y=1)'
         ) \%>\%
         mutate_if(is.character, as.factor) \%>\%
         mutate(CurrPain = factor(CurrPain),
         Neovascularization = factor(Neovascularization),
         Sex = factor(Sex))

enhanced_stripchart(PSFR ~ Side, data = TendonData)
enhanced_stripchart(HeelRaise ~ CurrPain, data = TendonData)
}
\references{
Ho K-Y, Baquet A, Chang Y-J, Chien L-C, Harty M, Bashford G, et al. (2019) Factors related to intra-tendinous morphology of Achilles tendon in runners. PLoS ONE 14(8): e0221183. https://doi.org/10.1371/journal.pone.0221183
}
\keyword{ANOVA}
\keyword{achilles}
\keyword{data}
\keyword{measures}
\keyword{repeated}
\keyword{tendon}
